Add non-working attempt at finding n_clusters using kmeans
This attempt simply iterates over possible n_clusters (from
MIN_CENTERS to MAX_CENTERS), and returns a KMeans model which gives
the highest score on the dataset.

This does not work - for pretty obvious reasons. When increasing the
number of clusters, the distance within a cluster will always decrease
(due to the number of data points within a cluster decreasing).
This means that this method always ends up with MAX_CENTERS being the
best choice of n_clusters.

Still committing it, to track my progress